Bump ScalaJS 0.6.8
add some workarounds for ShrinkSpecification

- https://github.com/rickynils/scalacheck/commit/f1ea446138ccac5a3c0
- https://github.com/rickynils/scalacheck/commit/155a054c61572724508
- https://travis-ci.org/rickynils/scalacheck/builds/106758308
