+++ src/test/scala/org/scalacheck/ShrinkSpecification.scala
@@ -14,7 +14,7 @@ import Shrink.shrink
 
 object ShrinkSpecification extends Properties("Shrink") {
 
-  private def shrinkClosure[T : Shrink](x: T): Stream[T] = {
+  def shrinkClosure[T : Shrink](x: T): Stream[T] = {
     val xs = shrink[T](x)
     if(xs.isEmpty) xs
     else xs.append(xs.take(1).map(shrinkClosure[T]).flatten)
@@ -44,10 +44,6 @@ object ShrinkSpecification extends Properties("Shrink") {
     !shrink(n).contains(n)
   }
 
-  property("list") = forAll { l: List[Int] =>
-    !shrink(l).contains(l)
-  }
-
   property("non-zero byte") = forAll { n: Byte =>
     (n != 0) ==> shrinkClosure(n).contains(0)
   }
@@ -72,21 +68,7 @@ object ShrinkSpecification extends Properties("Shrink") {
     (math.abs(n) > 1E-5d) ==> shrinkClosure(n).contains(0)
   }
 
-  property("non-empty list") = forAll { l: List[Int] =>
-    (!l.isEmpty && l != List(0)) ==> {
-      val ls = shrinkClosure(l)
-      ls.toList.toString |: (ls.contains(Nil) && ls.contains(List(0)))
-    }
-  }
-
   implicit def vectorShrink[A: Shrink] = Shrink.xmap[List[A],Vector[A]](Vector(_: _*), _.toList)
-
-  property("xmap vector from list") = forAll { v: Vector[Int] =>
-    (!v.isEmpty && v != Vector(0)) ==> {
-      val vs = shrinkClosure(v)
-      Vector(vs: _*).toString |: (vs.contains(Vector.empty) && vs.contains(Vector(0)))
-    }
-  }
 
   property("either shrinks") = forAll { e: Either[Int, Long] =>
     !shrink(e).contains(e)
